LayoutTests/platform/wpe/js/navigator-set-geolocation-expected.txt

Tests that navigator.geolocation cannot be shadowed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Strict mode
PASS navigator.geolocation = 1 threw exception TypeError: Attempted to assign to readonly property..
PASS navigator.geolocation is not 1
* Non-Strict mode
PASS navigator.geolocation = 1 did not throw exception.
PASS navigator.geolocation is not 1
PASS successfullyParsed is true

TEST COMPLETE
